+++ src/verifylib.rs
